pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "needle-haystack-ai"
version = "0.1.0"
description = "Needle RAG tools for Haystack"
authors = [
    "Onur Eken <m.onureken@gmail.com>",
    "Jan Heimes <heimesjan9@gmail.com>",
]
readme = "README.md"
packages = [{ include = "needle_haystack" }]
license = "MIT"
keywords = [
    "needle",
    "api",
    "retrieval-augmented generation",
    "rag",
    "information-retrieval",
    "artificial intelligence",
    "ai",
]
classifiers = [
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Development Status :: 4 - Beta",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
]

[tool.poetry.urls]
homepage = "https://needle-ai.com"
documentation = "https://docs.needle-ai.com"
repository = "https://github.com/JANHMS/needle-haystack"
issues = "https://github.com/JANHMS/needle-haystack/issues"

[tool.poetry.dependencies]
python = "^3.9"
haystack-ai = "^2.4.0"
needle-python = "^0.3.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.1"
pytest-mock = "^3.14.0"
pytest-cov = "^5.0.0"

[tool.poetry.group.lint.dependencies]
ruff = "0.5.6"
mypy = "1.11.1"

[tool.poetry.dev-dependencies]
pytest = "^8.2.2"
openai = "^1.39.0"
ipykernel = "^6.29.5"
ipywidgets = "^8.1.3"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
pre-commit = "^3.7.1"


[tool.ruff]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "D",  # pydocstyle
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long, handled by black
    "C901", # function is too complex
    "D206", # indent with spaces, not recommended with formatter
    "W191", # tab-indentation, not recommended with formatter
    "D203", # no-blank-line-before-class
    "D213", # multi-line-summary-second-line
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D"]

[tool.ruff.lint.isort]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
plugins = []
strict_optional = true
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = ["haystack", "haystack.utils"]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "--verbose"
filterwarnings = 'error'
xfail_strict = true
testpaths = ["tests"]
markers = ["integration: integration tests"]

[tool.coverage.run]
source = ["needle_haystack"]
branch = true

[tool.coverage.report]
precision = 2
exclude_lines = [
    'pragma: no cover',
    'raise NotImplementedError',
    'if TYPE_CHECKING:',
    '@overload',
]

[tool.coverage.xml]
output = "coverage.xml"
